requirements.txt
PyYAML
requests
tqdm
Pillow
pdf2image
boto3         # for Amazon Textract
openai        # for o4